Experimenting with i2c
I accidentally purchased an i2c nintendo controller thinking it was a
USB model. Then I thought about trying to convert it to USB. This is my
experiment reading the values using an i2c bus on a beaglebone black.

I took the code from a blog post and combined it with the knowledge from
other projects. Trying to understand i2c as implemented by nintendo
is quite tricky.

For the longest time I was actually doing the right thing, but still
XOR'ing the bytes assuming they were encrypted. Now I am successfully
able to read the bits for at least 2 buttons.